Crate s1_noisefloor_engine

Backbone engine for Sentinel-1 noise floor removal. Main interfaces are within the interface and postprocess modules

Modules

apply
Methods used to apply and subtract estimated noisefields

est_lp
Estimation method specific to linear programming method

estimate
Estimation procedures specific to linear method.

interface
Interface for applying the methods in specific cases.

parse
Module is responsible for extracting data directly from xml files.

postprocess
Routines for post processing the scene (e.g. multilooking, handling negative)

prep_lp
Top level module for preparing measurements for the linear programming implementation. This is everything up-to the actual program estimation

read_from_archive
Module for reading from sentinel-1 zip archive.