Function s1_noisefloor_engine::estimate::estimate_k_values

pub fn estimate_k_values(
    x: ArrayView2<f64>, 
    y: ArrayView2<f64>, 
    w: &[usize], 
    swath_bounds: &[&[SwathElem]], 
    mu: f64, 
    gamma: f64, 
    lambda_: &[f64], 
    _lambda2_: f64
) -> Array1<f64>

Estimates k values for square of image x and noise field y Assumes that x and y have already been divided by 10000.0 for scaling purposes.

Arguments

x square of input hv-SAR image

y ESA calibrated noise for hv image

w half-period in terms of row spacing

swath_bounds boundaries of subswaths in col indices