Module s1_noisefloor_engine::interface

Interface for applying the methods in specific cases.

Functions

linear_get_customscale_data
Applies the linear scaling method using custom user provided scales, k. Returns the values back in square intensity units.

linear_get_dualpol_data
Applies the lstsquares estimation method to retrieve scaling parameters, k, rescales the noise floor, y, and subtracts it from the image, x. Returns the values back in square intensity units. Parameters:

linear_get_raw_data
Returns the original cross pol, co pol, and noise field from the archive. Note that arrays are in linear units.

lp_get_customscale_data
Applies the power function noise floor obtained from linear programming, with parameters given by the user. Returns the values back in square intensity units.

lp_get_dualpol_data
Applies the linear programming method to restimate a noise floor based on the characteristics of the original image and the Returns the values back in square intensity units.