Module s1_noisefloor_engine::read_from_archive

Module for reading from sentinel-1 zip archive.

Structs

LpAttributes

SentinelFormatId

Enums

SentinelArchiveOutput

Functions

get_data_from_zip_path
Gets the original 16-bit image and noise field from a given path. path can either be the path to the zip archive, or an unzipped directory

get_id_prefix
Get the dateid prefix for the relevant files. ex: S1A_EW_GRDM_1SDH_20180902T164932_20180902T165032_023522_028FAA_5A8B.SAFE-report-20180902T190505.pdf Returns (Uppercase string, lowercase dateid, sentinelid[a or b])